add test data. update documentation. update temp file handling

+++ R/vs_fastq_trim.R
@@ -100,9 +100,11 @@ vs_fastq_trim <- function(fastq_input,
 
   # Set up cleanup of temporary files
   on.exit({
-    existing_files <- temp_files[file.exists(temp_files)]
-    if (length(existing_files) > 0) {
-      file.remove(existing_files)
+    if (length(temp_files) > 0 && is.character(temp_files)) {
+      existing_files <- temp_files[file.exists(temp_files)]
+      if (length(existing_files) > 0) {
+        file.remove(existing_files)
+      }
     }
   }, add = TRUE)
